requirements.txt
mypy==0.981
mypy-extensions==0.4.3
pyrsistent==0.18.0
PyYAML==6.0
tomli==2.0.1
types-PyYAML==6.0.12
typing_extensions==4.3.0
matplotlib==3.6.2
pandas==1.5.2
SQLAlchemy==2.0.9
psycopg2-binary==2.9.6